Parental Interview Discouraged
This is my PID. There are many like it, but this one is mine.
My PID is my best friend. It is my life. I must master it as I must master my life.
Without me, my PID is useless. Without my PID, I am useless. I must fire my PID true. I must shoot straighter than my enemy who is trying to kill me. I must shoot him before he shoots me. I will...
My PID and I know that what counts in war is not the rounds we fire, the noise of our burst, nor the smoke we make. We know that it is the hits that count. We will hit...
My PID is human, even as I, because it is my life. Thus, I will learn it as a brother. I will learn its weaknesses, its strength, its parts, its accessories, its sights and its barrel. I will keep my PID clean and ready, even as I am clean and ready. We will become part of each other. We will...
Before God, I swear this creed. My PID and I are the defenders of my country. We are the masters of our enemy. We are the saviors of my life.
So be it, until victory is Ameri